Change "ingress controller" to "Ingress"
"Ingress" and "ingress controller" are easily confused, since the Ingress is simply a set of instructions that the ingress controller uses to program an HTTP entrypoint. One does not create or modify an ingress controller after installation of the software into the cluster. One creates an Ingress, which is what the instructions on this page reflect.

These changes correct the language in this page of the docs.